Rename endVertex in EndCondition to vertex

+++ Pathfinding/src/main/java/pathfinding/FlexibleGraphExample.java
@@ -9,7 +9,7 @@ public class FlexibleGraphExample {
     public static void main(String[] args) {
         var graphRandomizer = FlexibleGraphRandomizer.<Character>builder()
                 .vertices(List.of('A', 'B', 'C', 'D', 'E'))
-                .weightFunction((source, destination) -> (double) Math.abs(source - destination))
+                .weightFunction((source, destination) -> Math.abs(source - destination))
                 .build();
 
         var graph = graphRandomizer.randomizeDirectedEdges();
